feat: add game registry

+++ server/src/main/java/de/thm/holdem/model/user/UserExtra.java
@@ -4,6 +4,8 @@ import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigInteger;
+
 @Data
 @Document(collection = "userextra")
 public class UserExtra {
@@ -11,7 +13,7 @@ public class UserExtra {
     @Id
     private String username;
     private String avatar;
-    private int bankroll;
+    private BigInteger bankroll;
 
     public UserExtra(String username) {
         this.username = username;
